show list on a fragment instead of the activity.
