+++ app/src/main/java/org/otfusion/caturday/ui/activities/FavoriteActivity.java
@@ -1,39 +1,22 @@
 package org.otfusion.caturday.ui.activities;
 
-import android.os.Bundle;
+import android.app.FragmentManager;
+import android.app.FragmentTransaction;
 import android.support.v7.widget.Toolbar;
-import android.view.ContextMenu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.ListView;
 
 import org.otfusion.caturday.R;
-import org.otfusion.caturday.common.model.Cat;
 import org.otfusion.caturday.ui.adapters.FavoriteCatAdapter;
+import org.otfusion.caturday.ui.fragments.BaseFragment;
+import org.otfusion.caturday.ui.fragments.FavoriteCatListFragment;
 import org.otfusion.caturday.ui.fragments.callbacks.FavoriteCallback;
-import org.otfusion.caturday.util.UIUtils;
-
-import java.util.List;
 
 import butterknife.Bind;
 
 public class FavoriteActivity extends CatActivity implements FavoriteCallback {
 
-    @Bind(R.id.favorite_list_view)
-    protected ListView mFavoriteCatsView;
-
     @Bind(R.id.favorite_toolbar)
     Toolbar mToolbar;
-
-    private FavoriteCatAdapter mFavoriteCatAdapter;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        refreshFavoriteCats();
-    }
 
     @Override
     protected int getContentLayoutId() {
@@ -42,18 +25,7 @@ public class FavoriteActivity extends CatActivity implements FavoriteCallback {
 
     @Override
     protected void loadUIContent() {
-        mFavoriteCatAdapter = new FavoriteCatAdapter();
-        mFavoriteCatsView.setAdapter(mFavoriteCatAdapter);
-        registerForContextMenu(mFavoriteCatsView);
         setSupportActionBar(mToolbar);
-
-        mFavoriteCatsView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-
-            }
-        });
-
         mToolbar.setNavigationIcon(getResources().getDrawable(R.drawable
                 .abc_ic_ab_back_mtrl_am_alpha));
         mToolbar.setNavigationOnClickListener(new View.OnClickListener() {
@@ -62,53 +34,27 @@ public class FavoriteActivity extends CatActivity implements FavoriteCallback {
                 onBackPressed();
             }
         });
+        showFragment();
     }
 
-    @Override
-    public boolean onContextItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_favorite_context_delete:
-                handleContextMenuDeleteOption(item);
-                return true;
-            case R.id.action_favorite_context_share:
-                UIUtils.showToast("Soon");
-                return true;
-            default:
-                return super.onContextItemSelected(item);
-        }
+    private void showFragment() {
+        FragmentManager fragmentManager = getFragmentManager();
+        FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+        BaseFragment fragment = getUsedFragment();
+        fragmentTransaction.add(R.id.fragment_container, fragment);
+        fragmentTransaction.commit();
     }
 
-    private void handleContextMenuDeleteOption(MenuItem item) {
-        AdapterView.AdapterContextMenuInfo menuInfo;
-        List<Cat> catFromsAdapter = getFavoriteCatAdapter().getCats();
-        menuInfo = getAdapterContextMenuInfo(item.getMenuInfo());
-        Cat favoritedCat = catFromsAdapter.get(menuInfo.position);
-        getCatService().deleteFromFavorites(favoritedCat);
-        refreshFavoriteCats();
-    }
-
-    @Override
-    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo
-            menuInfo) {
-        super.onCreateContextMenu(menu, v, menuInfo);
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.favorite_context_menu, menu);
-    }
-
-    private List<Cat> getFavoriteCats() {
-        return getCatService().getFavoriteCats();
-    }
-
-    private void refreshFavoriteCats() {
-        getFavoriteCatAdapter().updateCats(getFavoriteCats());
-    }
-
-    public FavoriteCatAdapter getFavoriteCatAdapter() {
-        return mFavoriteCatAdapter;
+    BaseFragment getUsedFragment() {
+        // should add logic here to check activity status/current fragment in case of being
+        // destroyed, so we can return the used fragment.
+        // or from a service?
+        return new FavoriteCatListFragment();
     }
 
     @Override
     public void showFavoritedCatImage() {
-        // show FavoritedCatImage
+        // TODO show FavoritedCatImage fragment
+        // the UsedFragment should be a ListFragment if not, do not display the image! ;)
     }
 }
