+++ app/src/main/res/layout/activity_favorite.xml
@@ -6,21 +6,28 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     tools:context="org.otfusion.caturday.ui.activities.FavoriteActivity">
 
+
+    <!--<fragment android:name="org.otfusion.caturday.ui.fragments.FavoriteCatListFragment"-->
+        <!--android:id="@+id/fragment_favorite_list"-->
+        <!--android:layout_width="match_parent"-->
+        <!--android:layout_height="match_parent"-->
+        <!--tools:layout="@layout/fragment_favorite_cat_list" />-->
+
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/fragment_layout">
+        android:id="@+id/fragment_container">
 
-        <ListView
-            android:id="@+id/favorite_list_view"
-            android:layout_width="fill_parent"
-            android:layout_height="fill_parent"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true"
-            android:layout_alignParentTop="true"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginRight="@dimen/activity_horizontal_margin"
-            android:layout_marginTop="?attr/actionBarSize" />
+        <!--<ListView-->
+            <!--android:id="@+id/favorite_list_view"-->
+            <!--android:layout_width="fill_parent"-->
+            <!--android:layout_height="fill_parent"-->
+            <!--android:layout_alignParentLeft="true"-->
+            <!--android:layout_alignParentStart="true"-->
+            <!--android:layout_alignParentTop="true"-->
+            <!--android:layout_marginLeft="@dimen/activity_horizontal_margin"-->
+            <!--android:layout_marginRight="@dimen/activity_horizontal_margin"-->
+            <!--android:layout_marginTop="?attr/actionBarSize" />-->
     </RelativeLayout>
 
     <android.support.v7.widget.Toolbar
